Add MockOdbGatewayService
Extract a common interface from ObdGatewayService and put it to AbstractGatewayService.
With this refactoring add the MockObdGatwayService which will be used when no BT
device or BT connectivity is present and help in debugging the application.

Also, run executeQueue in a different thread than the UI-thread in order to stop
"too much work on main thread" from appearing.