applied some threading love.
AbstractGatewayService.java: instead of (potentially) starting a new executeQueue thread every time queueJob is called, only start it _once_ in onCreate and more
 importantly have it interrupted by onDestroy. this means a new thread is started when service is created and asked nicely to stop when service is destroyed
(Mock)ObdGatewayService.java: try to handle the thread interruption by bailing out when thread is interrupted and catching InterruptedException (and re-throwing it),
no need for isQueueRunning flag anymore.
This resolves issue where multiple executeQueue threads are reading from stream at same time resulting in garbage out. in theory it should only have started new
thread when none is running and stopped when the queue was empty but this is wasn't always working. also no need to stop the thread when queue is empty it will
happily sit and wait for it to be refilled again so we can just have one thread.